+++ CHANGES.md
@@ -1,6 +1,7 @@
 HEAD
 ----
 
+* Allow `Slop#present?` to accept multiple arguments
 * Added `:all_accept_arguments` to Slop configuration options, this saves
   having to specify that every option takes an argument
 * Added `Slop#to_struct` for building new classes from options